// src/store/modules/kmr/mutation-types.ts
/*
 * @Description: app mutations type
 * @Author: Rongxis
 * @Date: 2020-12-23 10:25:37
 * @LastEditors: Rongxis
 * @LastEditTime: 2020-12-23 10:38:43
 */

export enum KmrMutationTypes {
  SET_MODAL_RECT = 'SET_MODAL_RECT',
  SET_CLIENT_RECT = 'SET_CLIENT_RECT',
  SET_RECORD_STEP = 'SET_RECORD_STEP',
  SET_RECORD_STATE = 'SET_RECORD_STATE',
  SET_RECORD_HIT_SELF = 'SET_RECORD_HIT_SELF',
  SET_CLIENT_FREEZING = 'SET_CLIENT_FREEZING',
}