// src/store/modules/kmr/action-types.ts
/*
 * @Description: app actions type
 * @Author: Rongxis
 * @Date: 2020-12-23 10:25:37
 * @LastEditors: Rongxis
 * @LastEditTime: 2020-12-23 14:02:25
 */

export enum KmrActionTypes {
  ACTION_SET_CLIENT_RECT = 'ACTION_SET_CLIENT_RECT',
  ACTION_SET_CLIENT_FREEZING = 'ACTION_SET_CLIENT_FREEZING',
  ACTION_SET_MODAL_RECT = 'ACTION_SET_MODAL_RECT',
  ACTION_SET_RECORD_START = 'ACTION_SET_RECORD_START',
  ACTION_SET_RECORD_PAUSE = 'ACTION_SET_RECORD_PAUSE',
  ACTION_SET_RECORD_STOP = 'ACTION_SET_RECORD_STOP',
  ACTION_SET_RECORD_STEP = 'ACTION_SET_RECORD_STEP',
  ACTION_ADD_RECORD_HIT_SELF = 'ACTION_ADD_RECORD_HIT_SELF',
}